missense_kinase_toolkit/databases/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "versioningit"]
build-backend = "setuptools.build_meta"

[project]
name = "mkt-databases"
dynamic = ["version"]
authors = [
  { name="Jess White", email="jessica.white@choderalab.org" },
]
description = "Module in missense-kinase-toolkit for querying databases"
readme = "README.md"
requires-python = ">=3.9,<3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]

dependencies = [
    "mkt.schema",
    "gitpython==3.1.43",
    "pydantic>=2.7.4",
    "tqdm>=4.66.3",
    "pandas>=2,<3",
    "requests>=2.28.1,<3",
    "requests-cache>=0.9.7,<1",
    "bravado>=11.0.3",
    "janitor>=0.1.1",
    "beautifulsoup4>=4.12.3",
    "ipykernel>=6.29.4",
    "jupyter>=1.0.0",
    "biopython>=1.83",
    "bokeh>=3.4.1",
    "seaborn>=0.13.2",
    "scipy>=1.13.1",
    "matplotlib-venn>=1.1.1",
    "pyhmmer>=0.10.15",
    "biotite>=1.0.1",
    "openpyxl>=3.1.5",
    "upsetplot>=0.9.0",
]

[project.optional-dependencies]
dev = ["black==24.3.0", "flake8==7.0.0", "sphinx==7.3.7", "sphinx-rtd-theme==2.0.0"]
test = ["pytest==8.0.1", "pytest-runner==6.0.1"]

[project.scripts]
generate_kinaseinfo_objects = "mkt.databases.cli.generate_kinaseinfo_objects:main"

[project.urls]
"Homepage" = "https://github.com/choderalab/missense-kinase-toolkit"
"Bug Tracker" = "https://github.com/choderalab/missense-kinase-toolkit/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.versioningit]
default-version = "1+unknown"

[tool.versioningit.format]
distance = "{base_version}+{distance}.{vcs}{rev}"
dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"
distance-dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"

[tool.versioningit.vcs]
method = "git"
match = ["*"]
default-tag = "0.0.0"
